+++ @utahdts/utah-design-system/index.js
@@ -25,7 +25,7 @@ export { ConfirmationChildren } from './react/components/buttons/ConfirmationBut
 export { InitialChildren } from './react/components/buttons/ConfirmationButton/InitialChildren';
 export { DocumentationTemplate } from './react/components/templates/DocumentationTemplate';
 export { ExternalLink } from './react/components/navigation/ExternalLink';
-export { FileInput } from './react/components/forms/FileInput';
+export { FileInput } from './react/components/forms/FileInput/FileInput';
 export { FileTemplate } from './react/components/forms/FileInput/FileTemplate';
 export { FooterAgencyInformation } from './react/components/footer/FooterAgencyInformation';
 export { FooterAgencyInformationColumn } from './react/components/footer/FooterAgencyInformationColumn';
